% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{abssumdiff}
\alias{abssumdiff}
\title{Finding the absolute sum of the differences}
\usage{
abssumdiff(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector of numeric values}

\item{na.rm}{single logical should missing values be removed}
}
\value{
A single value data.
}
\description{
Finding the absolute sum of the differences
called by \code{segmentation}.
}
\examples{
   tmp1 <- c(1,3,2,6,4,5,3,9,10)
   abssumdiff(x = tmp1)
}
\keyword{internal}
