% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{CirVar}
\alias{CirVar}
\title{CirVar function}
\usage{
CirVar(rotation, na.rm = TRUE)
}
\arguments{
\item{na.rm}{single logical should missing values be removed}

\item{x}{vector of numeric values}
}
\value{
A single value data.
}
\description{
Called by \code{segmentation}.
}
\examples{
   tmp1 <- c(1,3,2,6,4,5,3,9,10)
   CirVar(tmp1)
}
\keyword{internal}
