% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImport.R
\name{degrees}
\alias{degrees}
\title{Extract rotation time series}
\usage{
degrees(x)
}
\arguments{
\item{x}{data output from get.intervals}
}
\value{
The degrees (rotation) data.
}
\description{
Extract data relating to the rotation component.
}
\details{
Called by \code{dataImport}.
Note: the "+ 1" has been removed from the original implementation.
}
\examples{
   d1 = matrix(c(100, 101, -0.79, -0.86, -0.17, -0.14, 0.53, 0.46), 
        nrow = 2, ncol = 4)
   colnames(d1) = c("timestamp", "x", "y", "z")
   degrees(x = d1)
}
\keyword{internal}
