% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsGENEA.R
\name{features}
\alias{features}
\title{Get features of a GENEA rpart fit}
\usage{
features(fit)
}
\arguments{
\item{fit}{GENEA rpart object}
}
\value{
Character vector naming features selected by GENEA rpart fit.
}
\description{
Shows the response features specified by a GENEA (rpart) object.
}
\examples{
data(trainingFit)
features(fit = trainingFit)
}
\keyword{internal}
