% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepCounter.R
\name{find_peaks}
\alias{find_peaks}
\title{Find_Peaks}
\usage{
find_peaks(x, m = 3)
}
\arguments{
\item{x}{timeseries signal}

\item{m}{The number of points either side of the peak to required to be a peak.}
}
\description{
Peak detector function
}
\details{
Finds the peaks and valleys within the signal passed to the function.
}
\keyword{Internal}
