% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\docType{package}
\name{GENEAclassify-package}
\alias{GENEAclassify-package}
\title{Classification of accelorometer data}
\description{
This package provides tools to perform the segmentation
and classification of GENEActiv accelorometer data. The high frequency
time series data is split into segments based on activity changepoints.
An \pkg{rpart} fit is trained against known activities at each segment.
This fit can then then be used to guess behaviours from test data when
activity at each time point has not been reported. This allows detailed
behaviour profiles to be created for the wearer.
}
\keyword{package}
