% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{GENEAcount}
\alias{GENEAcount}
\title{Count Steps}
\usage{
GENEAcount(x, ...)
}
\arguments{
\item{x}{vector}

\item{\dots}{other arguments to be swallowed}
}
\value{
single numeric
}
\description{
Step count is the number of step lengths divided by 2.
}
\examples{
   x1 <- c(20, 15, 10)
   GENEAcount(x1)
   x2 <- c(300, 255, 111)
   GENEAcount(x2)
}
\keyword{internal}
