% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{GENEAdistance}
\alias{GENEAdistance}
\title{Find the mean distance between two steps detected}
\usage{
GENEAdistance(x, ...)
}
\arguments{
\item{x}{vector}

\item{\dots}{other arguments to be swallowed}
}
\value{
single numeric
}
\description{
Average Step wavelength over the segment
}
\examples{
   x1 <- c(20, 15, 10)
   GENEAdistance(x1)
   x2 <- c(300, 255, 111)
   GENEAdistance(x2)
}
\keyword{internal}
