% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{GENEAskew}
\alias{GENEAskew}
\title{Find average Amplitude of Steps}
\usage{
GENEAskew(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{vector}

\item{\dots}{other arguments to be swallowed}
}
\value{
single numeric
}
\description{
Step amplitude finds the average size of the step found over the segment
}
\examples{
   x1 <- c(20, 15, 10)
   GENEAcount(x1)
   x2 <- c(300, 255, 111)
   GENEAcount(x2)
}
\keyword{internal}
