% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{GENEAwavelength}
\alias{GENEAwavelength}
\title{Find average Wavelength of Steps}
\usage{
GENEAwavelength(x, ...)
}
\arguments{
\item{x}{vector}

\item{\dots}{other arguments to be swallowed}
}
\value{
single numeric
}
\description{
Average Step wavelength over the segment
}
\examples{
   x1 <- c(20, 15, 10)
   GENEAwavelength(x1)
   x2 <- c(300, 255, 111)
   GENEAwavelength(x2)
}
\keyword{internal}
