% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsGENEA.R
\name{inGroup}
\alias{inGroup}
\title{Which values are in a new group?}
\usage{
inGroup(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
logical vector with length x
}
\description{
Choose elements at each step.
Used by \code{\link{features}} to identify unique components of splits table.
}
\examples{
GENEAclassify:::inGroup(x = rep(-1, 5))
x1 <- c(1, 1, 1, 2, 2, 2, 2, 3, 3)
x1t <- GENEAclassify:::inGroup(x = x1)
x1[x1t]
}
\keyword{internal}
