% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsGENEA.R
\name{levels.GENEA}
\alias{levels.GENEA}
\title{Get levels of a GENEA rpart response}
\usage{
\method{levels}{GENEA}(x)
}
\arguments{
\item{x}{GENEA rpart object}
}
\value{
character vector naming levels classified by GENEA rpart fit
}
\description{
Shows the levels used in a GENEA (rpart) fit object.
}
\examples{
data(trainingFit)
levels(x = trainingFit)
}
\keyword{internal}
