% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{meandiff}
\alias{meandiff}
\title{Finding the mean of the differences}
\usage{
meandiff(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector of numeric values}

\item{na.rm}{Remove the na values in the vector}
}
\value{
A single value data.
}
\description{
Finding the mean of the differences
Called by \code{segmentation}.
}
\examples{
   tmp1 <- c(1,3,2,6,4,5,3,9,10)
   meandiff(x = tmp1)
}
\keyword{internal}
