% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{MeanDir}
\alias{MeanDir}
\title{MeanDir function}
\usage{
MeanDir(rotation, na.rm = T)
}
\arguments{
\item{na.rm}{single logical should missing values be removed}

\item{x}{vector of numeric values}
}
\value{
A single value data.
}
\description{
Called by \code{segmentation}.
}
\examples{
   tmp1 <- c(1,3,2,6,4,5,3,9,10)
   MeanDir(tmp1)
}
\keyword{internal}
