% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepCounter.R
\name{runmean}
\alias{runmean}
\title{calculate the running mean}
\usage{
runmean(x, k = 2)
}
\arguments{
\item{x}{numeric vector}

\item{k}{single integer window size}
}
\value{
numeric vector, debiased
}
\description{
calculate the running mean, bumping the window against the ends of the segment
borrowed in part from \pkg{caTools}
}
\examples{
    x1 <- c(-1, 1, 2, 2, 3, 3, 3, 4, 5)
    runmean(x1, k = 3)
}
\keyword{internal}
