% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{sumdiff}
\alias{sumdiff}
\title{Finding the sum of the differences}
\usage{
sumdiff(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector of numeric values}

\item{na.rm}{single logical should missing values be removed}
}
\value{
A single value data.
}
\description{
Adding in a feedback call sumdiff that finds the difference of a function and then sums this difference. 
Called by \code{segmentation}.
}
\examples{
   tmp1 <- c(1,3,2,6,4,5,3,9,10)
   sumdiff(x = tmp1)
}
\keyword{internal}
