% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGENEAmodel.R
\docType{data}
\name{TrainingData}
\alias{TrainingData}
\title{Example Training Data set}
\format{Manually classified segmented data that can be used to build a classification model.}
\source{
Manually classified by Actvinsights.
}
\description{
A manually classified training data set provided with the package.
}
\examples{
data(TrainingData)
class(TrainingData)
}
\seealso{
\code{"\link[=levels.GENEA]{levels}"} \code{\link{features}}
}
\keyword{datasets}
