% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGENEAmodel.R
\docType{data}
\name{trainingFit}
\alias{trainingFit}
\title{Example classification tree}
\format{
An \code{rpart} object with class \code{GENEA} containing the decision 
tree information required for prediction.
}
\source{
Output of \code{createGENEAmodel} on experimental training data.
}
\description{
rpart object declaring the decision tree for the classification of GENEActiv bin data into typical groups.
}
\examples{
data(trainingFit)
class(trainingFit)
levels(trainingFit)
features(trainingFit)
plot(trainingFit)
text(trainingFit, cex = 0.5)
}
\seealso{
\code{"\link[=levels.GENEA]{levels}"} \code{\link{features}}
}
\keyword{datasets}
