% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImport.R
\name{updown}
\alias{updown}
\title{Extract data relating to the up/down component}
\usage{
updown(x)
}
\arguments{
\item{x}{data output from \code{get.intervals}}
}
\value{
The up/down vertical elevation data (y-axis)
}
\description{
Extract up/down time series
}
\details{
Extract code from \code{positionals} to perform data conversion to up/down time series.
Input is expected to be result of \code{get.intervals}.
}
\examples{
    d1 = matrix(c(100, 101, -0.79, -0.86, -0.17, -0.14, 0.53, 0.46), 
        nrow = 2, ncol = 4)
    colnames(d1) = c("timestamp", "x", "y", "z")
    updown(x = d1)
}
\keyword{internal}
