% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{GENEAskew}
\alias{GENEAskew}
\title{Skewness, a measure of centredness}
\usage{
GENEAskew(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{single logical should missing values be removed}

\item{\dots}{other arguments to be swallowed}
}
\value{
single numeric
}
\description{
Many datasets do not have the highest density 
of measurements in the middle of the distribution of the data.
Skewness is a measure of the asymmetry of a probability distribution.
}
\examples{
GENEAskew(1:10)
GENEAskew((1:10)^2)
GENEAskew((1:10)^0.5)
}
\keyword{internal}
